requirements.txt
fastapi[standard]==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
alembic